Fixed costOfOrder, was int now is float

+++ src/Customer.java
@@ -4,11 +4,11 @@ public class Customer {
 	private int arrival;
 	private String id;
 	private int timeToPrepare;
-	private int costOfOrder;
+	private float costOfOrder;
 	private int patience;
 	private Boolean satisfied;
 	
-	public Customer(int arrival, String id, int timeToPrepare, int costOfOrder, int patience) {
+	public Customer(int arrival, String id, int timeToPrepare, float costOfOrder, int patience) {
 		this.arrival = arrival;
 		this.id = id;
 		this.timeToPrepare = timeToPrepare;
@@ -28,7 +28,7 @@ public class Customer {
 		return timeToPrepare;
 	}
 	
-	public int getCostOfOrder() {
+	public float getCostOfOrder() {
 		return costOfOrder;
 	}
